Revert "DELETE all folders to start over again. Decided not to use sample code from MVVM Hacker News example because some parts I don't understand yet, and I think thiswould make adding onto the sample code difficult. The APIRequest and NewsViewModel are the parts I don't understand."
This reverts commit 6b38196f7a66fb685d33e49c57c0f32aacd46883.

Reason for reverting is the Hacker News Algolia API does not get as many articles as the Hacker News Firebase API used in previous MVVM demo sample code app. I requested 100 and it returned much less than that, about 37 articles.

Other problems with Algolia API are it has a request limit, its articles are not as up-to-date as actual Hacker News website, and articles like Show HN do not display a webpage.