feat: Share sheet to share a news article's title and URL to another app.
E.g. by chat app.

Mad an ActivityViewController in order to show the share sheet in iOS < 16. Made separate classes for sheet selected objects to fix bug in SwiftUI, where on first tap the share sheet was blank. Needed to use StateObject instead of State according to answers on StackOverflow.