feat: Save or unsave article from details screen.
As a user, being able to save an article or remove it in DetailView is convenient, because if I go back to the browsing news list the article may be replaced by more recent ones.

Previously, a user could only save an article in browse news list. That means I save an article based on limited info, like if its title and upvotes look interesting in the list, or I go to DetailView, then go back if I like the article. That can be inconvenient because the article may get replaced when the list gets updated.

DetailView now accepts 2 extra arguments, Item and FavoriteItem.

The Story struct now saves the entire news Item to send to DetailView.

The FavoriteStory struct now sends the FavoriteItem to DetailView.

DetailView handles 3 cases:
- It was navigated from favorites list, in which case it shows the heart filled button
- It was navigated from browse news list, and the Item is already saved to favorites, based on having same title and poster. In which case it shows heart filled button
- It was navigated from browse news list, and the Item was not saved to favorites. In which case it shows heart unfilled button